merge
todo: ui and nu
