+++ app/assets/javascripts/discourse/app/components/modal/create-account.js
@@ -21,6 +21,7 @@ import { findAll } from "discourse/models/login-method";
 import User from "discourse/models/user";
 import discourseDebounce from "discourse-common/lib/debounce";
 import discourseComputed, {
+  bind,
   observes,
   on,
 } from "discourse-common/utils/decorators";
@@ -53,6 +54,90 @@ export default Component.extend(
     authOptions: alias("model.authOptions"),
     accountEmail: alias("model.accountEmail"),
     accountUsername: alias("model.accountUsername"),
+
+    didInsertElement() {
+      this._super(...arguments);
+
+      if (cookie("email")) {
+        this.set("model.accountEmail", cookie("email"));
+      }
+
+      let userTextFields = document.getElementsByClassName("user-fields")[0];
+
+      if (userTextFields) {
+        userTextFields =
+          userTextFields.getElementsByClassName("ember-text-field");
+      }
+
+      if (userTextFields) {
+        for (let element of userTextFields) {
+          element.addEventListener("focus", this.userInputFocus);
+          element.addEventListener("focusout", this.userInputFocusOut);
+        }
+      }
+
+      this.element.addEventListener("keydown", this.actionOnEnter);
+      this.element.addEventListener("click", this.selectKitFocus);
+    },
+
+    willDestroyElement() {
+      this._super(...arguments);
+
+      this.element.removeEventListener("keydown", this.actionOnEnter);
+      this.element.removeEventListener("click", this.selectKitFocus);
+
+      let userTextFields = document.getElementsByClassName("user-fields")[0];
+
+      if (userTextFields) {
+        userTextFields =
+          userTextFields.getElementsByClassName("ember-text-field");
+      }
+
+      if (userTextFields) {
+        for (let element of userTextFields) {
+          element.removeEventListener("focus", this.userInputFocus);
+          element.removeEventListener("focusout", this.userInputFocusOut);
+        }
+      }
+    },
+
+    // used for animating the label inside of inputs
+    userInputFocus(event) {
+      const userField = event.target.parentElement.parentElement;
+      if (!userField.classList.contains("value-entered")) {
+        userField.classList.toggle("value-entered");
+      }
+    },
+
+    // used for animating the label inside of inputs
+    userInputFocusOut(event) {
+      const userField = event.target.parentElement.parentElement;
+      if (
+        event.target.value.length === 0 &&
+        userField.classList.contains("value-entered")
+      ) {
+        userField.classList.toggle("value-entered");
+      }
+    },
+
+    @bind
+    actionOnEnter(event) {
+      if (!this.submitDisabled && event.key === "Enter") {
+        event.preventDefault();
+        event.stopPropagation();
+        this.createAccount();
+        return false;
+      }
+    },
+
+    @bind
+    selectKitFocus(event) {
+      const target = document.getElementById(event.target.getAttribute("for"));
+      if (target?.classList.contains("select-kit")) {
+        event.preventDefault();
+        target.querySelector(".select-kit-header").click();
+      }
+    },
 
     @discourseComputed(
       "hasAuthOptions",
