+++ app/assets/stylesheets/desktop/login.scss
@@ -206,12 +206,11 @@
 // create account
 // modal only
 .d-modal.create-account {
-  .create-account-body {
-    min-width: 100%;
-  }
   .modal-body {
+    min-width: 100%;
     overflow: hidden;
   }
+
   .has-alt-auth .create-account-form {
     display: grid;
     grid-template-columns: 60% 40%;
